+++ components/Workspaces/RepositoryStatCard.tsx
@@ -61,14 +61,18 @@ const getPillChart = (total?: number, loading?: boolean) => {
   }
 
   if (total > 7) {
-    return <Pill icon={<ArrowTrendingUpIcon color="green" className="w-4 h-4" />} text="High" color="green" />;
+    return (
+      <Pill icon={<ArrowTrendingUpIcon color="green" className="w-6 h-6 lg:w-4 lg:h-4" />} text="High" color="green" />
+    );
   }
 
   if (total >= 4 && total <= 7) {
-    return <Pill icon={<MinusSmallIcon color="black" className="w-4 h-4" />} text="Medium" color="yellow" />;
+    return (
+      <Pill icon={<MinusSmallIcon color="black" className="w-6 h-6 lg:w-4 lg:h-4" />} text="Medium" color="yellow" />
+    );
   }
 
-  return <Pill icon={<ArrowTrendingDownIcon color="red" className="w-4 h-4" />} text="Low" color="red" />;
+  return <Pill icon={<ArrowTrendingDownIcon color="red" className="w-6 h-6 lg:w-4 lg:h-4" />} text="Low" color="red" />;
 };
 
 const EmptyState = ({ type, hasError }: { type: CardType; hasError: boolean }) => {
@@ -104,12 +108,12 @@ export const RepositoryStatCard = ({ stats, type, isLoading, hasError }: Reposit
   const loadEmptyState = isLoading || hasError || !stats;
 
   return (
-    <Card className="w-80 max-w-xs h-32 max-h-32">
+    <Card className="w-full lg:w-80 lg:max-w-xs lg:h-32 lg:max-h-32">
       {loadEmptyState ? (
         <EmptyState type={type} hasError={hasError} />
       ) : (
         <table className="grid gap-4 p-2">
-          <caption className="flex items-center gap-1.5 text-xs">
+          <caption className="flex items-center gap-1.5 lg:text-xs">
             {getIcon(type)}
             <span>{titles[type]}</span>
           </caption>
@@ -119,13 +123,15 @@ export const RepositoryStatCard = ({ stats, type, isLoading, hasError }: Reposit
               .map(([stat, value]) => {
                 return (
                   <tr key={stat} className="flex flex-col">
-                    <th scope="row" className="capitalize font-medium text-sm text-light-slate-11 text-left">
+                    <th scope="row" className="capitalize font-medium text-lg lg:text-sm text-light-slate-11 text-left">
                       {stat.replace("_", " ")}
                     </th>
                     {stat === "activity_ratio" ? (
-                      <td className="text-black semi-bold text-2xl">{getPillChart(Math.round(value), isLoading)}</td>
+                      <td className="text-black semi-bold text-3xl lg:text-2xl">
+                        {getPillChart(Math.round(value), isLoading)}
+                      </td>
                     ) : (
-                      <td className="semi-bold text-2xl" title={`${value}`}>
+                      <td className="semi-bold text-3xl lg:text-2xl" title={`${value}`}>
                         {stat === "velocity" ? `${Math.round(value)}d` : humanizeNumber(value, "abbreviation")}
                       </td>
                     )}